arm: DT: MSM8996-mdss: Rework rotator clock names
On k4.4 we are using the SDE Rotator regardless of being MDSS or
SDE enabled.
The new SDE Rotator driver is made for MSM8998, which has got
different clocks, so QC changed the names to reflect those.
I find it a good thing, talking about 8998, but they screwed up
the good old MSM8996!

Rework rotator clock names to get in line with 8998 names and
initialize the driver just fine.